Fix compiler warnings
src/hotplug_macosx.c:291:32: warning:
      implicit conversion loses integer precision: 'unsigned long' to 'UInt32'
      (aka 'unsigned int') [-Wshorten-64-to-32]
  ...driverBundle->m_vendorId = strtoul(stringBuffer, NULL, 16);
                              ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/hotplug_macosx.c:296:33: warning:
      implicit conversion loses integer precision: 'unsigned long' to 'UInt32'
      (aka 'unsigned int') [-Wshorten-64-to-32]
  ...driverBundle->m_productId = strtoul(stringBuffer, NULL, 16);
                               ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/hotplug_macosx.c
@@ -288,12 +288,12 @@ static HPDriverVector HPDriversGetFromDirectory(const char *driverBundlePath)
 
 				CFStringGetCString(strValue, stringBuffer, sizeof stringBuffer,
 					kCFStringEncodingUTF8);
-				driverBundle->m_vendorId = strtoul(stringBuffer, NULL, 16);
+				driverBundle->m_vendorId = (unsigned int)strtoul(stringBuffer, NULL, 16);
 
 				strValue = CFArrayGetValueAtIndex(productArray, j);
 				CFStringGetCString(strValue, stringBuffer, sizeof stringBuffer,
 					kCFStringEncodingUTF8);
-				driverBundle->m_productId = strtoul(stringBuffer, NULL, 16);
+				driverBundle->m_productId = (unsigned int)strtoul(stringBuffer, NULL, 16);
 
 				strValue = CFArrayGetValueAtIndex(friendlyNameArray, j);
 				CFStringGetCString(strValue, stringBuffer, sizeof stringBuffer,
